♻️ REFACTOR: SSE 관련 클래스는 web의 하위 폴더 sse에 위치
