+++ src/main/java/cloudcomputinginha/demo/web/sse/SseService.java
@@ -1,8 +1,8 @@
-package cloudcomputinginha.demo.service.notification;
+package cloudcomputinginha.demo.web.sse;
 
 import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
 
-public interface NotificationSseService {
+public interface SseService {
     public SseEmitter subscribe(Long memberId, String lastEventId);
 
     public void sendNotification(SseEmitter emitter, String emitterId, String eventId, Object data);
